+++ app/Http/Controllers/Menu/Contractor/QualificationController.php
@@ -8,7 +8,8 @@ use App\Models\Qualification;
 use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Support\Str;
-use Intervention\Image\Facades\Image;
+use Intervention\Image\ImageManager;
+use Intervention\Image\Drivers\Gd\Driver;
 
 class QualificationController extends Controller
 {
@@ -99,17 +100,24 @@ class QualificationController extends Controller
             'issue_date' => isset($request->issue_date) ? Carbon::parse($request->issue_date)->startOfDay() : null,
             'expiry_date' => isset($request->expiry_date) ? Carbon::parse($request->expiry_date)->startOfDay() : null
         ]);
-        // Set image if it exists.
-        if (isset($request->image)) {
-            // Create the new image.
+        // Check the request data for the required file.
+        if ($request->hasFile('image')) {
+            // Set the uploaded file.
             $image = $request->file('image');
-            $filename = Str::slug($selected_user->getFullNameAttribute() . '-' . $request->title) . '.' . $image->getClientOriginalExtension(); 
+            // Set the new file name.
+            $filename = Str::slug($selected_user->getFullNameAttribute() . '-' . $request->title) . '.' . $image->getClientOriginalExtension();
+            // Set the new file location.
             $location = storage_path('app/public/images/qualifications/' . $filename);
-            Image::make($image)->orientate()->save($location);
+            // Create new manager instance with desired driver.
+            $manager = new ImageManager(new Driver());
+            // Read image from filesystem
+            $image = $manager->read($image);
+            // Encoding jpeg data
+            $image->toJpeg(80)->save($location);
             // Update the selected model instance.
-            $new_qualification->update([[
+            $selected_user->update([
                 'image_path' => 'storage/images/qualifications/' . $filename
-            ]]);
+            ]);
         }
         // Return a redirect to the index route.
         return redirect()
@@ -172,19 +180,25 @@ class QualificationController extends Controller
             'issue_date' => isset($request->issue_date) ? Carbon::parse($request->issue_date)->startOfDay() : null,
             'expiry_date' => isset($request->expiry_date) ? Carbon::parse($request->expiry_date)->startOfDay() : null
         ]);
-        // Qualification image.
-        if (isset($request->image)){
-            // Delete the old image from storage.
-            if ($selected_qualification->image_path != null) {
-                if (file_exists(public_path($selected_qualification->image_path))) {
-                    unlink(public_path($selected_qualification->image_path));
-                }
+        // Check the request data for the required file.
+        if ($request->hasFile('image')) {
+            // Check if the file path value is not null and file exists on the server.
+            if ($selected_qualification->image_path != null && file_exists(public_path($selected_qualification->image_path))) {
+                // Delete the file from the server.
+                unlink(public_path($selected_qualification->image_path));
             }
-            // Create the new image.
+            // Set the uploaded file.
             $image = $request->file('image');
-            $filename = Str::slug($selected_qualification->staff->getFullNameAttribute() . '-' . $request->title) . '.' . $image->getClientOriginalExtension(); 
+            // Set the new file name.
+            $filename = Str::slug($selected_qualification->staff->getFullNameAttribute() . '-' . $request->title) . '.' . $image->getClientOriginalExtension();
+            // Set the new file location.
             $location = storage_path('app/public/images/qualifications/' . $filename);
-            Image::make($image)->orientate()->save($location);
+            // Create new manager instance with desired driver.
+            $manager = new ImageManager(new Driver());
+            // Read image from filesystem
+            $image = $manager->read($image);
+            // Encoding jpeg data
+            $image->toJpeg(80)->save($location);
             // Update the selected model instance.
             $selected_qualification->update([
                 'image_path' => 'storage/images/qualifications/' . $filename
@@ -204,14 +218,16 @@ class QualificationController extends Controller
      */
     public function destroy($id)
     {
+        // ?????????????????????
+
+        // This might be broken due to calling request and sending it as a get parameter.
+
         // Find the required model instance.
         $selected_qualification = Qualification::findOrFail($id);
-        // Delete the images relationship instances.
-        if ($selected_qualification->image_path != null) {
-            // Delete the file from storage.
-            if (file_exists(public_path($selected_qualification->image_path))) {
-                unlink(public_path($selected_qualification->image_path));
-            }
+        // Check if the file path value is not null and file exists on the server.
+        if ($selected_qualification->image_path != null && file_exists(public_path($selected_qualification->image_path))) {
+            // Delete the file from the server.
+            unlink(public_path($selected_qualification->image_path));
         }
         // Delete the selected model instance.
         $selected_qualification->delete();
